# openapi.yaml
openapi: 3.1.0
info:
  title: users_service API
  version: 0.3.0
  description: |
    Контрактный API users_service.
    Авторизация выполняется через Bearer access token (JWT, issuer=auth_service).
servers:
  - url: http://localhost:8000

tags:
  - name: health
  - name: admin
  - name: user
  - name: parent

paths:
  /healthz:
    get:
      tags: [health]
      summary: Проверка состояния
      operationId: healthCheck
      responses:
        '200':
          description: Сервис доступен
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok

  /v1/admin/users:
    post:
      tags: [admin]
      summary: Создать пользователя
      operationId: adminCreateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

    get:
      tags: [admin]
      summary: Список пользователей
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/Role'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }

  /v1/admin/users/{user_id}:
    get:
      tags: [admin]
      summary: Получить пользователя по id
      operationId: adminGetUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

    patch:
      tags: [admin]
      summary: Обновить профиль пользователя
      operationId: adminUpdateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /v1/admin/users/{user_id}/roles:
    post:
      tags: [admin]
      summary: Назначить роль пользователю
      operationId: adminAssignRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: Роль назначена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [admin]
      summary: Снять роль у пользователя
      operationId: adminRevokeRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - in: query
          name: role
          required: true
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Роль снята
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/admin/users/{user_id}/block:
    post:
      tags: [admin]
      summary: Заблокировать пользователя
      operationId: adminBlockUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Пользователь заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/admin/users/{user_id}/unblock:
    post:
      tags: [admin]
      summary: Разблокировать пользователя
      operationId: adminUnblockUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Пользователь разблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/admin/users/{user_id}/archive:
    post:
      tags: [admin]
      summary: Архивировать пользователя
      operationId: adminArchiveUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Пользователь архивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/admin/users/{user_id}/restore:
    post:
      tags: [admin]
      summary: Восстановить архивированного пользователя
      operationId: adminRestoreUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Пользователь восстановлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/admin/links:
    get:
      tags: [admin]
      summary: Список связей родитель-ученик
      operationId: adminListLinks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: parent_id
          schema:
            type: string
        - in: query
          name: student_id
          schema:
            type: string
      responses:
        '200':
          description: Список связей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentStudentLinkListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }

    post:
      tags: [admin]
      summary: Создать связь родитель-ученик
      operationId: adminCreateLink
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParentStudentLinkRequest'
      responses:
        '201':
          description: Связь создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentStudentLinkResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /v1/admin/links/{link_id}:
    delete:
      tags: [admin]
      summary: Удалить связь родитель-ученик
      operationId: adminRemoveLink
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LinkId'
      responses:
        '200':
          description: Связь удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentStudentLinkResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/user/me:
    get:
      tags: [user]
      summary: Получить свой профиль
      operationId: userGetMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Собственный профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }

    patch:
      tags: [user]
      summary: Обновить свой профиль
      operationId: userUpdateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /v1/parent/me/students:
    get:
      tags: [parent]
      summary: Список связанных учеников
      operationId: parentListStudents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ученики родителя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
    LinkId:
      name: link_id
      in: path
      required: true
      schema:
        type: string

  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    AccessDenied:
      description: Доступ запрещен
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Конфликт или нарушение инварианта
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Ошибка валидации
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  schemas:
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }

    Role:
      type: string
      enum: [admin, teacher, parent, student]

    UserStatus:
      type: string
      enum: [active, blocked, archived]

    CreateUserRequest:
      type: object
      required: [email, display_name, roles]
      properties:
        user_id:
          type: string
          nullable: true
        email:
          type: string
        display_name:
          type: string
        phone:
          type: string
          nullable: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'

    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true

    AssignRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/Role'

    UserResponse:
      type: object
      required:
        [user_id, email, display_name, status, roles, created_at, updated_at, version]
      properties:
        user_id:
          type: string
        email:
          type: string
        display_name:
          type: string
        phone:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer

    UserListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'

    CreateParentStudentLinkRequest:
      type: object
      required: [parent_id, student_id]
      properties:
        link_id:
          type: string
          nullable: true
        parent_id:
          type: string
        student_id:
          type: string
        note:
          type: string
          nullable: true

    ParentStudentLinkStatus:
      type: string
      enum: [requested, active, removed]

    ParentStudentLinkResponse:
      type: object
      required:
        [link_id, parent_id, student_id, status, created_at, updated_at, version]
      properties:
        link_id:
          type: string
        parent_id:
          type: string
        student_id:
          type: string
        status:
          $ref: '#/components/schemas/ParentStudentLinkStatus'
        note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer

    ParentStudentLinkListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ParentStudentLinkResponse'